Moved test into one file and fixed example to be clearer
